# api/assets/openapi.json
{
	"openapi": "3.0.0",
	"servers": [
		{
			"url": "https://api.copycord.com/api/v{version}",
			"description": "Official copycord instance",
			"variables": {
				"version": {
					"default": "9",
					"enum": ["8", "9"]
				}
			}
		}
	],
	"info": {
		"description": "copycord is a free open source selfhostable discord compatible chat, voice and video platform",
		"version": "1.0.0",
		"title": "copycord HTTP API Routes",
		"termsOfService": "",
		"contact": {
			"name": "copycord"
		},
		"license": {
			"name": "AGPLV3",
			"url": "https://www.gnu.org/licenses/agpl-3.0.en.html"
		}
	},
...
	"paths": {
...
		"/guilds/templates/{code}": {
			"get": {
				"description": "",
				"responses": {
					"default": {
						"description": "not documented"
					}
				},
				"parameters": [
					{
						"name": "code",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						},
						"description": "code"
					}
				],
				"tags": ["guilds"]
			},
			"post": {
				"description": "",
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GuildTemplateCreateSchema"
							}
						}
					}
				},
				"responses": {
					"default": {
						"description": "not documented"
					}
				},
				"parameters": [
					{
						"name": "code",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						},
						"description": "code"
					}
				],
				"tags": ["guilds"]
			}
		},
...
		"/guilds/{guild_id}/templates/{code}": {
			"delete": {
				"description": "##### Requires the ``MANAGE_GUILD`` permission\n",
				"responses": {
					"default": {
						"description": "not documented"
					}
				},
				"parameters": [
					{
						"name": "guild_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						},
						"description": "guild_id"
					},
					{
						"name": "code",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						},
						"description": "code"
					}
				],
				"tags": ["guilds"]
			},
			"put": {
				"description": "##### Requires the ``MANAGE_GUILD`` permission\n",
				"responses": {
					"default": {
						"description": "not documented"
					}
				},
				"parameters": [
					{
						"name": "guild_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						},
						"description": "guild_id"
					},
					{
						"name": "code",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						},
						"description": "code"
					}
				],
				"tags": ["guilds"]
			},
			"patch": {
				"description": "##### Requires the ``MANAGE_GUILD`` permission\n",
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TemplateModifySchema"
							}
						}
					}
				},
				"responses": {
					"default": {
						"description": "not documented"
					}
				},
				"parameters": [
					{
						"name": "guild_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						},
						"description": "guild_id"
					},
					{
						"name": "code",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						},
						"description": "code"
					}
				],
				"tags": ["guilds"]
			}
		},
...
	},
...
	"components": {
		"schemas": {
...
			"GuildTemplateCreateSchema": {
				"type": "object",
				"properties": {
					"name": {
						"type": "string"
					},
					"avatar": {
						"type": "string",
						"nullable": true
					}
				},
				"required": ["name"]
			},
...
			"TemplateModifySchema": {
				"type": "object",
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					}
				},
				"required": ["name"]
			},
...
		},
...
		"securitySchemes": {
			"Token": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			}
		},
...
	}
}